requirements.txt
numpy
matplotlib
plotly
clarifai-grpc
